Basic quick turn functionality has been implemented.
If you check the [VRTK_SDKManager] in the Hierarchy, you will see a new script attached to the child object [VRTK_SDKSetups]. The script is called QuickTurn. It works by manipulating the Transform.rotation of the object. There is a serialized field that allows you to adjust how many degrees the quick turn will turn you. Look at the code for more information.

Problems:
Since the code uses the horizontal axis to "smoothly" turn the player, we turn right now while rotating. We need to figure out some axis that is not being used by any of the motion controls to get this to work.

We also might need to fine tune the degree amount of the quick turn.