Contributed the KeyItem script and made my own test scene
KeyItem works like this.
It makes key items purple when they are not grabbed.
When an item is dropped, it teleports back to its spawn after a certain period of time

The script just needs to be put on an object that has a VRTK_Interactable  Object script attached to it. There seems to be some slight bugs. I think it's because sometimes I'm using different VRTK interactions than the ones I'm checking for.