Inspectable and InspectableDNA scripts
Inspectable allows the player to be able to manipulate their object by rotating in the certain planes of motion as well as scaling them.

InspectableDNA is a derived class that allows the player to manipulate the object in the same manner as if the object were to have the Insepectable script but with the added functionality of being able to combine or separate strands. The strands will ideally start off as separate.