Implemented the prototype of the DNA puzzle.
Problems with the current implementation...

The DNA prefabs I created have rotations and positions that might cause problems down the line. I don't have any idea how to reset the transform of a GameObject.

We might have to switch out the text mesh since we are working in VRTK. There's probably something better.

I used a mouse click to cycle through the bases of the DNA.

We might need to add some sort of scene transition after we're done solving the DNA.

I don't know how this scales with bigger DNA.