Vibrio Isolation Code
So the Vibrio Q-Tip, am object with a Swab component, can "swab" an object with the "PetriDish" tag and PetriDish script attached.

I think the problem might be that we have to predominantly work with ints to color the texture but the petri dish model is so tiny that it gets easily covered. But enlargening the petri dish still resulted in the same problem.

There's a strange bug. The swab doesn't quite "draw" on the petri dish but it does draw on a flat cylinder? Check my scene to experiment further.

Basically the code checks an array of pixels in the texture to know when to stop. This is why the texture size is 64 * 64. Setting it higher is not recommended. If about 75% of the array is a desired color, it stops swabbing. The desired color is set for each petri dish. The swab's color is compared against the dish's color. (this is a way of checking if the sample is wrong)