Properly aligned DNA half segments with TextMeshPro objects
This is a minor push for fixing some features of the DNA. I readjusted the pivots of the DNA segments to facilitate an future attempt to implement the ability for the player to "spin" the DNA in place. I also am going to try to figure out the proper way to implement a click. Once that's done, my plan is to put this in the wrap up room along with the petri dish. Then experiment with the shader graph. The reason being is that the shader graph seems to mess up the objects that currently have no materials and the petri dish is one of those.

The changes made for the DNA
=============================
-Changed the pivots of the model
-Realigned parts of the actual model
-TextMeshPro is added
-Change the TextMesh objects to TextMeshPro objects (this will have crisper text)
-Fixed some bugs that resulted from the changes...