Finally "fixed" the DNA and put it in the Wrap-Up scene
I basically taught myself about delegates and event handlers in order to implement "clicking". Basically, if you put your hand near a strand, it'll turn orange. If you touch it, it turns green, changing base. Once you match all the pairs, the halves slowly merge together. The highlights and the base symbols will disappear.

I could not implement rotation. Because rotation in VRTK seems to require basically a solid connection between what is being roated and the axis or joint it rotates around. In my opinion, rotation seems a bit senseless.

Some things to comment on:
The highlighting is a bit wonky in the wrap-up scene. Specifically, the Near Touch highlights don't work. I think it might be because the player hands don't actually have the InteractNearTouch script attached. If we don't want to attach that, that's fine.

I also didn't put in the shader graph yet because I want to wait until everyone is done  putting stuff in the wrap-up scene before I start experimenting with that.