Logic Board Pathway Updates
More updates to the logic board to have a flow for each of the tests. IGS requires evidence to continue, which is to be printed once virus placed into the microscope. DNA Test requires player to screenshot the matching percentage and place on board. Each is connected to its own button (Phase 1 complete -> IGS unlocks, Phase 2 -> DNA, Phase 3 -> Submit).