Changed Opening Line Progression
Modified ToolCenter.cs and AIVoice.cs to ensure that VIRA finishes all opening lines before the player can interact with the tool station and progress through the game. To implement this change, the Ai Voice value for the Tool Center script attached to Tool Center Destroyer / Spawner object in the WrapUp scene was replaced with AI Audio Source prefab (original value was null).