Egg and Flashlight Interactions
Adjusted the two scripts to work based on trigger enter of flashlight cone to avoid pushing the egg with light. Also discovered the issue of why VRTK physics-based grabs would not work - all positions were locked. Fixing this allowed the fixed joint grab to work, which does not child the object (meaning it can still collide with objects in the player's hand that are children).