Flask Collection + Alpha Egg Cut
Added the collection of infected cells in the holographic display. Each object has its own VirtualCell script that will save the respawn location for incorrect additions (working on functionality). When added, the flask adds onto the count for the test. Also dumping the collection of files changed for the progress on Egg cut in Brendan's Scene.