Snap Rotation & Testing Room Scale
Set up a rotatable hex tile using a placeholder made in Blender. Rotation on the object used is very important, needs to be zeroed out and initially in the correct rotation to rotate along the proper axis. Scaling scene created and adjusted according to Game Design doc on unity scale.